+++ packages/labextension/src/index.ts
@@ -158,11 +158,11 @@ const extension: JupyterLabPlugin<void> = {
         }
       });
 
-      console.log("Jupytext: adding command=" + command + " with rank=" + rank);
-      palette.addItem({ command, rank, category: "Jupytext" });
+      console.log("Jupytext: adding command=" + command + " with rank=" + (rank + 1));
+      palette.addItem({ command, rank: rank + 2, category: "Jupytext" });
     });
 
-    // Jupytext README
+    // Jupytext's documentation
     palette.addItem({
       args: {
         text: "Jupytext Reference",
@@ -170,7 +170,17 @@ const extension: JupyterLabPlugin<void> = {
       },
       command: "help:open",
       category: "Jupytext",
-      rank: JUPYTEXT_FORMATS.length + 1
+      rank: 0
+    });
+
+    palette.addItem({
+      args: {
+        text: "Jupytext FAQ",
+        url: "https://jupytext.readthedocs.io/en/latest/faq.html"
+      },
+      command: "help:open",
+      category: "Jupytext",
+      rank: 1
     });
 
     // Metadata in text representation    
@@ -230,7 +240,7 @@ const extension: JupyterLabPlugin<void> = {
       }
     });
 
-    palette.addItem({ command: "jupytext_metadata", rank: JUPYTEXT_FORMATS.length + 1, category: "Jupytext" });
+    palette.addItem({ command: "jupytext_metadata", rank: JUPYTEXT_FORMATS.length + 3, category: "Jupytext" });
   }
 };